all current test has added, but some of them are still falling

+++ Drill4Net.Target.Common/src/Misc/Finalizer.cs
@@ -8,7 +8,7 @@ namespace Drill4Net.Target.Common
 
         /***************************************************/
 
-        public Finalizer(ushort len)
+        public Finalizer(int len)
         {
             _ms = new MemoryStream(new byte[len]);
         }
